Migrate from makefile to CMake (#11)
* Introduce switch for stripping symbols by linker
* Add option to use LLVM's LLD as linker
* Add option for saving temp files during build process
The specific use case for keeping these temp files is to be able
to perform analysis and comparison on the produced assembly files
of the kernel. For example, diff-ing these files with previous
versions and seeing what a code change does to the produced assembly.
* Remove obsolete BLOCK specification
* Remove lld-variant of linker script
With the removal of BLOCK specification from linker.ld, the two
variants of the linker script are now identical. Therefore there is
no reason to keep them separate anymore.
Although GCC is not fully functional yet, this compiler flag is
blatantly not supported and therefore should be removed.
* Remove separate GCC and LLVM makefiles