Enhance CI workflow to add analysis files to private repo (#17)
To facilitate binary analysis of the kernel code, temporarily until a proper web service is developed for this task, the
build by-products that are valuable for binary analysis such as the assembly code are uploaded to a private repo. This
allows for viewing a diff of the effects of PR code changes using GitHub.