Rename src/boot.asm to src/boot.S so g++ recognizes it as ASM
When cross-compiling using GNU toolchain, invoking make results in
"src/boot.asm: linker input file unused because linking not done".
The end result is that no object file gets produced for boot.asm,
which means the linking will fail.

The issue boils down to the fact that g++ is unable to read and
compile this assembly source code to an object file so it gets
linked at a later time.

I found that this issue can be resolved by:
- Adding `set(CMAKE_ASM_FLAGS "${CFLAGS} -x assembler-with-cpp")`
to the block in CMakeLists.txt that is relevant for GNU toolchain.
- Changing the extension of boot.asm to boot.S

Given that the first option means that we are setting the source
language of the GNU toolchain to be assembler-with-cpp rather than
just cpp, I believe the second option (renaming boot.asm) is a less
invasive option. I have tested both and they seem to work (not 100%
sure if they result in identical binary file). I am choosing to just
rename boot.asm to boot.S to resolve this issue.